+++ lib/active_storage_validations/content_type_validator.rb
@@ -2,30 +2,42 @@
 
 module ActiveStorageValidations
   class ContentTypeValidator < ActiveModel::EachValidator # :nodoc:
-    def validate_each(record, attribute, _value)
-      return true if !record.send(attribute).attached? || types.empty?
+    include OptionProcUnfolding
 
+    AVAILABLE_CHECKS = %i[with in].freeze
+    
+    def validate_each(record, attribute, _value)
+      return true unless record.send(attribute).attached?
+
+      types = authorized_types(record)
+      return true if types.empty?
+      
       files = Array.wrap(record.send(attribute))
 
-      errors_options = { authorized_types: types_to_human_format }
+      errors_options = { authorized_types: types_to_human_format(types) }
       errors_options[:message] = options[:message] if options[:message].present?
 
       files.each do |file|
-        next if is_valid?(file)
+        next if is_valid?(file, types)
 
         errors_options[:content_type] = content_type(file)
-        record.errors.add(attribute, :content_type_invalid, errors_options)
+        record.errors.add(attribute, :content_type_invalid, **errors_options)
         break
       end
     end
 
-    def types
-      (Array.wrap(options[:with]) + Array.wrap(options[:in])).compact.map do |type|
-        Mime[type] || type
+    def authorized_types(record)
+      flat_options = unfold_procs(record, self.options, AVAILABLE_CHECKS)
+      (Array.wrap(flat_options[:with]) + Array.wrap(flat_options[:in])).compact.map do |type|
+        if type.is_a?(Regexp)
+          type
+        else
+          Marcel::MimeType.for(declared_type: type.to_s, extension: type.to_s)
+        end
       end
     end
 
-    def types_to_human_format
+    def types_to_human_format(types)
       types
         .map { |type| type.to_s.split('/').last.upcase }
         .join(', ')
@@ -35,11 +47,10 @@ module ActiveStorageValidations
       file.blob.present? && file.blob.content_type
     end
 
-    def is_valid?(file)
-      if options[:with].is_a?(Regexp)
-        options[:with].match?(content_type(file).to_s)
-      else
-        content_type(file).in?(types)
+    def is_valid?(file, types)
+      file_type = content_type(file)
+      types.any? do |type|
+        type == file_type || (type.is_a?(Regexp) && type.match?(file_type.to_s))
       end
     end
   end