+++ lib/active_storage_validations/limit_validator.rb
@@ -2,29 +2,31 @@
 
 module ActiveStorageValidations
   class LimitValidator < ActiveModel::EachValidator # :nodoc:
+    include OptionProcUnfolding
+
     AVAILABLE_CHECKS = %i[max min].freeze
 
     def check_validity!
       return true if AVAILABLE_CHECKS.any? { |argument| options.key?(argument) }
-
       raise ArgumentError, 'You must pass either :max or :min to the validator'
     end
 
     def validate_each(record, attribute, _)
       files = Array.wrap(record.send(attribute)).compact.uniq
-      errors_options = { min: options[:min], max: options[:max] }
+      flat_options = unfold_procs(record, self.options, AVAILABLE_CHECKS)
+      errors_options = { min: flat_options[:min], max: flat_options[:max] }
 
-      return true if files_count_valid?(files.count)
-      record.errors.add(attribute, options[:message].presence || :limit_out_of_range, errors_options)
+      return true if files_count_valid?(files.count, flat_options)
+      record.errors.add(attribute, options[:message].presence || :limit_out_of_range, **errors_options)
     end
 
-    def files_count_valid?(count)
-      if options[:max].present? && options[:min].present?
-        count >= options[:min] && count <= options[:max]
-      elsif options[:max].present?
-        count <= options[:max]
-      elsif options[:min].present?
-        count >= options[:min]
+    def files_count_valid?(count, flat_options)
+      if flat_options[:max].present? && flat_options[:min].present?
+        count >= flat_options[:min] && count <= flat_options[:max]
+      elsif flat_options[:max].present?
+        count <= flat_options[:max]
+      elsif flat_options[:min].present?
+        count >= flat_options[:min]
       end
     end
   end