+++ lib/active_storage_validations/matchers/content_type_validator_matcher.rb
@@ -29,29 +29,39 @@ module ActiveStorageValidations
 
       def matches?(subject)
         @subject = subject.is_a?(Class) ? subject.new : subject
-        allowed_types_allowed? && rejected_types_rejected?
+        responds_to_methods && allowed_types_allowed? && rejected_types_rejected?
       end
 
       def failure_message
-        <<~MESSAGE
-          Expected #{@attribute_name}
-            
-            Accept content types: #{allowed_types.join(", ")}
-              #{accepted_types_and_failures}
+        message = ["Expected #{@attribute_name}"]
 
-            Reject content types: #{rejected_types.join(", ")}
-              #{rejected_types_and_failures}
-        MESSAGE
+        if @allowed_types
+          message << "Accept content types: #{allowed_types.join(", ")}"
+          message << "#{@missing_allowed_types.join(", ")} were rejected"
+        end
+
+        if @rejected_types
+          message << "Reject content types: #{rejected_types.join(", ")}"
+          message << "#{@missing_rejected_types.join(", ")} were accepted"
+        end
+
+        message.join("\n")
       end
 
       protected
+
+      def responds_to_methods
+        @subject.respond_to?(@attribute_name) &&
+          @subject.public_send(@attribute_name).respond_to?(:attach) &&
+          @subject.public_send(@attribute_name).respond_to?(:detach)
+      end
 
       def allowed_types
         @allowed_types || []
       end
 
       def rejected_types
-        @rejected_types || (Mime::LOOKUP.keys - allowed_types)
+        @rejected_types || []
       end
 
       def allowed_types_allowed?
@@ -62,22 +72,6 @@ module ActiveStorageValidations
       def rejected_types_rejected?
         @missing_rejected_types ||= rejected_types.select { |type| type_allowed?(type) }
         @missing_rejected_types.none?
-      end
-
-      def accepted_types_and_failures
-        if @missing_allowed_types.present?
-          "#{@missing_allowed_types.join(", ")} were rejected."
-        else
-          "All were accepted successfully."
-        end
-      end
-
-      def rejected_types_and_failures
-        if @missing_rejected_types.present?
-          "#{@missing_rejected_types.join(", ")} were accepted."
-        else
-          "All were rejected successfully."
-        end
       end
 
       def type_allowed?(type)