+++ lib/active_storage_validations/metadata.rb
@@ -1,9 +1,32 @@
 module ActiveStorageValidations
   class Metadata
+    class InvalidImageError < StandardError; end
+
     attr_reader :file
 
     def initialize(file)
+      require_image_processor
       @file = file
+    end
+
+    def image_processor
+      Rails.application.config.active_storage.variant_processor
+    end
+
+    def exception_class
+      if image_processor == :vips && defined?(Vips)
+        Vips::Error
+      elsif defined?(MiniMagick)
+        MiniMagick::Error
+      end
+    end
+
+    def require_image_processor
+      if image_processor == :vips
+        require 'vips' unless defined?(Vips)
+      else
+        require 'mini_magick' unless defined?(MiniMagick)
+      end
     end
 
     def metadata
@@ -14,6 +37,16 @@ module ActiveStorageValidations
           { width: image.width, height: image.height }
         end
       end
+    rescue InvalidImageError
+      logger.info "Skipping image analysis because ImageMagick or Vips doesn't support the file"
+      {}
+    end
+
+    def valid?
+      read_image
+      true
+    rescue InvalidImageError
+      false
     end
 
     private
@@ -21,7 +54,16 @@ module ActiveStorageValidations
     def read_image
       is_string = file.is_a?(String)
       if is_string || file.is_a?(ActiveStorage::Blob)
-        blob = is_string ? ActiveStorage::Blob.find_signed(file) : file
+        blob =
+          if is_string
+            if Rails.gem_version < Gem::Version.new('6.1.0')
+              ActiveStorage::Blob.find_signed(file)
+            else
+              ActiveStorage::Blob.find_signed!(file)
+            end
+          else
+            file
+          end
 
         tempfile = Tempfile.new(["ActiveStorage-#{blob.id}-", blob.filename.extension_with_delimiter])
         tempfile.binmode
@@ -33,29 +75,50 @@ module ActiveStorageValidations
         tempfile.flush
         tempfile.rewind
 
-        image = MiniMagick::Image.new(tempfile.path)
+        image = if image_processor == :vips && defined?(Vips) && Vips::get_suffixes.include?(File.extname(tempfile.path).downcase)
+                  Vips::Image.new_from_file(tempfile.path)
+                elsif defined?(MiniMagick)
+                  MiniMagick::Image.new(tempfile.path)
+                end
       else
-        image = MiniMagick::Image.new(read_file_path)
+        file_path = read_file_path
+        image = if image_processor == :vips && defined?(Vips) && Vips::get_suffixes.include?(File.extname(file_path).downcase)
+                  Vips::Image.new_from_file(file_path)
+                elsif defined?(MiniMagick)
+                  MiniMagick::Image.new(file_path)
+                end
       end
 
-      if image.valid?
-        yield image
-      else
-        logger.info "Skipping image analysis because ImageMagick doesn't support the file"
-        {}
-      end
-    rescue LoadError
-      logger.info "Skipping image analysis because the mini_magick gem isn't installed"
+
+      raise InvalidImageError unless valid_image?(image)
+      yield image if block_given?
+    rescue LoadError, NameError
+      logger.info "Skipping image analysis because the mini_magick or ruby-vips gem isn't installed"
       {}
-    rescue MiniMagick::Error => error
-      logger.error "Skipping image analysis due to an ImageMagick error: #{error.message}"
+    rescue exception_class => error
+      logger.error "Skipping image analysis due to an #{exception_class.name.split('::').map(&:downcase).join(' ').capitalize} error: #{error.message}"
       {}
     ensure
       image = nil
     end
 
+    def valid_image?(image)
+      return false unless image
+
+      image_processor == :vips && image.is_a?(Vips::Image) ? image.avg : image.valid?
+    rescue exception_class
+      false
+    end
+
     def rotated_image?(image)
-      %w[ RightTop LeftBottom ].include?(image["%[orientation]"])
+      if image_processor == :vips && image.is_a?(Vips::Image)
+        image.get('exif-ifd0-Orientation').include?('Right-top') ||
+          image.get('exif-ifd0-Orientation').include?('Left-bottom')
+      else
+        %w[ RightTop LeftBottom ].include?(image["%[orientation]"])
+      end
+    rescue exception_class # field "exif-ifd0-Orientation" not found
+      false
     end
 
     def read_file_path
@@ -63,7 +126,18 @@ module ActiveStorageValidations
       when ActionDispatch::Http::UploadedFile, Rack::Test::UploadedFile
         file.path
       when Hash
-        File.open(file.fetch(:io)).path
+        io = file.fetch(:io)
+        if io.is_a?(StringIO)
+          tempfile = Tempfile.new([File.basename(file[:filename], '.*'), File.extname(file[:filename])])
+          tempfile.binmode
+          IO.copy_stream(io, tempfile)
+          io.rewind
+          tempfile.flush
+          tempfile.rewind
+          tempfile.path
+        else
+          File.open(io).path
+        end
       else
         raise "Something wrong with params."
       end