+++ test/dummy/db/migrate/20180828144610_create_users.rb
@@ -1,9 +1,0 @@
-class CreateUsers < ActiveRecord::Migration[5.2]
-  def change
-    create_table :users do |t|
-      t.string :name
-
-      t.timestamps
-    end
-  end
-end