+++ test/dummy/db/schema.rb
@@ -1,63 +1,68 @@
-# This file is auto-generated from the current state of the database. Instead
-# of editing this file, please use the migrations feature of Active Record to
-# incrementally modify your database, and then regenerate this schema definition.
-#
-# This file is the source Rails uses to define your schema when running `rails
-# db:schema:load`. When creating a new database, `rails db:schema:load` tends to
-# be faster and is potentially less error prone than running all of your
-# migrations from scratch. Old migrations may fail to apply correctly if those
-# migrations use external dependencies or application code.
-#
-# It's strongly recommended that you check this file into your version control system.
+# frozen_string_literal: true
 
-ActiveRecord::Schema.define(version: 2019_10_17_145230) do
-
-  create_table "active_storage_attachments", force: :cascade do |t|
-    t.string "name", null: false
-    t.string "record_type", null: false
-    t.integer "record_id", null: false
-    t.integer "blob_id", null: false
-    t.datetime "created_at", null: false
-    t.index ["blob_id"], name: "index_active_storage_attachments_on_blob_id"
-    t.index ["record_type", "record_id", "name", "blob_id"], name: "index_active_storage_attachments_uniqueness", unique: true
+ActiveRecord::Schema.define do
+  create_table :active_storage_attachments, force: :cascade do |t|
+    t.string :name, null: false
+    t.string :record_type, null: false
+    t.integer :record_id, null: false
+    t.integer :blob_id, null: false
+    t.datetime :created_at, null: false
+    t.index %i[blob_id], name: :index_active_storage_attachments_on_blob_id
+    t.index %i[record_type record_id name blob_id], name: :index_active_storage_attachments_uniqueness, unique: true
   end
 
-  create_table "active_storage_blobs", force: :cascade do |t|
-    t.string "key", null: false
-    t.string "filename", null: false
-    t.string "content_type"
-    t.text "metadata"
-    t.bigint "byte_size", null: false
-    t.string "checksum", null: false
-    t.datetime "created_at", null: false
-    t.index ["key"], name: "index_active_storage_blobs_on_key", unique: true
+  create_table :active_storage_blobs, force: :cascade do |t|
+    t.string :key, null: false
+    t.string :filename, null: false
+    t.string :content_type
+    t.text :metadata
+    t.bigint :byte_size, null: false
+    t.string :checksum, null: false
+    t.datetime :created_at, null: false
+    t.index %i[key], name: :index_active_storage_blobs_on_key, unique: true
+
+    if Rails.gem_version >= Gem::Version.new('6.1.0')
+      t.string :service_name, null: false
+    end
   end
 
-  create_table "limit_attachments", force: :cascade do |t|
-    t.string "name"
+  if Rails.gem_version >= Gem::Version.new('6.1.0')
+    create_table :active_storage_variant_records, force: :cascade do |t|
+      t.bigint :blob_id, null: false
+      t.string :variation_digest, null: false
+      t.index %i[blob_id variation_digest], name: :index_active_storage_variant_records_uniqueness, unique: true
+    end
   end
 
-  create_table "only_images", force: :cascade do |t|
-    t.datetime "created_at", precision: 6, null: false
-    t.datetime "updated_at", precision: 6, null: false
+  create_table :documents, force: :cascade do |t|
+    t.datetime :created_at, precision: 6, null: false
+    t.datetime :updated_at, precision: 6, null: false
   end
 
-  create_table "projects", force: :cascade do |t|
-    t.string "title"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+  create_table :limit_attachments, force: :cascade do |t|
+    t.string :name
   end
 
-  create_table "ratio_models", force: :cascade do |t|
-    t.string "name"
-    t.datetime "created_at", precision: 6, null: false
-    t.datetime "updated_at", precision: 6, null: false
+  create_table :only_images, force: :cascade do |t|
+    t.datetime :created_at, precision: 6, null: false
+    t.datetime :updated_at, precision: 6, null: false
   end
 
-  create_table "users", force: :cascade do |t|
-    t.string "name"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+  create_table :projects, force: :cascade do |t|
+    t.string :title
+    t.datetime :created_at, null: false
+    t.datetime :updated_at, null: false
   end
 
+  create_table :ratio_models, force: :cascade do |t|
+    t.string :name
+    t.datetime :created_at, precision: 6, null: false
+    t.datetime :updated_at, precision: 6, null: false
+  end
+
+  create_table :users, force: :cascade do |t|
+    t.string :name
+    t.datetime :created_at, null: false
+    t.datetime :updated_at, null: false
+  end
 end