+++ test/matchers/attached_validator_matcher_test.rb
@@ -14,7 +14,7 @@ class ActiveStorageValidations::Matchers::AttachedValidatorMatcher::Test < Activ
     refute matcher.matches?(User)
   end
 
-  test 'unkown attached when providing class' do
+  test 'unknown attached when providing class' do
     matcher = ActiveStorageValidations::Matchers::AttachedValidatorMatcher.new(:non_existing)
     refute matcher.matches?(User)
   end
@@ -29,8 +29,38 @@ class ActiveStorageValidations::Matchers::AttachedValidatorMatcher::Test < Activ
     refute matcher.matches?(User.new)
   end
 
-  test 'unkown attached when providing instance' do
+  test 'unknown attached when providing instance' do
     matcher = ActiveStorageValidations::Matchers::AttachedValidatorMatcher.new(:non_existing)
     refute matcher.matches?(User.new)
+  end
+
+  test 'positive match with valid conditional validation' do
+    matcher = ActiveStorageValidations::Matchers::AttachedValidatorMatcher.new(:conditional_image)
+    assert matcher.matches?(User.new(name: 'Foo'))
+  end
+
+  test 'negative match with invalid conditional validation' do
+    matcher = ActiveStorageValidations::Matchers::AttachedValidatorMatcher.new(:conditional_image)
+    refute matcher.matches?(User.new)
+  end
+
+  test 'positive match when providing instance with attachment' do
+    matcher = ActiveStorageValidations::Matchers::AttachedValidatorMatcher.new(:avatar)
+    user = User.new
+    user.avatar.attach(io: Tempfile.new('.'), filename: 'image.png', content_type: 'image/png')
+    user.proc_avatar.attach(io: Tempfile.new('.'), filename: 'image.png', content_type: 'image/png')
+    assert matcher.matches?(user)
+  end
+
+  test 'positive match when providing persisted instance with attachment' do
+    matcher = ActiveStorageValidations::Matchers::AttachedValidatorMatcher.new(:avatar)
+    user = User.create!(
+      name: 'Pietje',
+      avatar: { io: Tempfile.new('.'), filename: 'image.png', content_type: 'image/png' },
+      photos: [{ io: Tempfile.new('.'), filename: 'image.png', content_type: 'image/png' }],
+      proc_avatar: { io: Tempfile.new('.'), filename: 'image.png', content_type: 'image/png' },
+      proc_photos: [{ io: Tempfile.new('.'), filename: 'image.png', content_type: 'image/png' }]
+    )
+    assert matcher.matches?(user)
   end
 end